fix some haha

+++ admin/index.php
@@ -11,17 +11,19 @@ require_once "../templates/header.php";
             <div class="input-container">
                 <input id="song-name" class="input" name="song-name" type="text" placeholder=" " required />
                 <div class="cut"></div>
-                <label for="song-name" class="placeholder">song name</label>
+                <label for="song-name" class="placeholder">name</label>
             </div>
             <div id="song-warning" class="warning"></div>
             <div class="input-container">
                 <input id="song-image" class="input" name="song-image" type="file" accept=".jpg, .jpeg, .png" placeholder=" " required />
-                <label for="song-image" class="placeholder">song image</label>
+                <div class="cut"></div>
+                <label for="song-image" class="placeholder">image</label>
             </div>
             <div id="image-warning" class="warning"></div>
             <div class="input-container">
-                <input id="song-file" name="song-file" accept=".mp3" type="file" placeholder=" " required />
-                <label for="song-file" class="placeholder">song file</label>
+                <input id="song-file" name="song-file" accept=".mp3, .m4a" class="input" type="file" placeholder=" " required />
+                <div class="cut"></div>
+                <label for="song-file" class="placeholder">song</label>
             </div>
             <div id="file-warning" class="warning"></div>
             <div class="input-container">
@@ -38,27 +40,31 @@ require_once "../templates/header.php";
                     }
                     ?>
                 </select>
+                <div class="cut"></div>
+                <label for="song-file" class="placeholder">singer</label>
             </div>
             <div id="select-warning" class="warning"></div>
             <button type="text" class="submit" id="song-submit-btn">Add a song</button>
         </form>
         <div id="edit-song-list">
-            <div class="edit-song">
-                <h4>dsad</h4>
-                <select id="edit-singer" class="input" type="select" required>
-                    <option value="">Select a singer</option>
-                    <?php
-                    require_once "../data/mysql-connection.php";
-                    $query = "SELECT singer_id, singer_name FROM singers;";
-                    $result = $conn->query($query);
-                    if ($result && $result->num_rows > 0) {
-                        while ($row = $result->fetch_assoc()) {
-                            echo "<option value='$row[singer_id]'>$row[singer_name]</option>";
-                        }
-                    }
-                    ?>
-                </select>
-            </div>
+            <?php require_once "../data/mysql-connection.php";
+            $query = "SELECT s.s_id, s.singer_id, sg.singer_name, s.s_name FROM songs s LEFT OUTER JOIN singers sg
+            ON s.singer_id = sg.singer_id
+            ORDER BY s_id;";
+            $result = $conn->query($query);
+            if ($result && $result->num_rows > 0) {
+                while ($row = $result->fetch_assoc()) {
+                    echo
+                    "<div class='edit-song'>
+                        <h4>$row[s_name]</h4>
+                        <div class='update-delete-icon flex' editFor='$row[s_id]'>
+                            <h5 class='active'>$row[singer_name]</h5>
+                            <i class='fa-solid fa-pen-to-square edit'></i>
+                            <i class='fa-solid fa-trash warning del'></i>
+                        </div>
+                    </div>";
+                }
+            } ?>
         </div>
 
     </div>
@@ -68,13 +74,12 @@ require_once "../templates/header.php";
             <div class="input-container">
                 <input id="singer-name" class="input" type="text" placeholder=" " />
                 <div class="cut"></div>
-                <label for="singer-name" class="placeholder">singer name</label>
+                <label for="singer-name" class="placeholder">singer</label>
             </div>
             <div id="singer-warning" class="warning"></div>
             <button type="text" id="singer-submit-btn" type="button" class="submit">Add new singer</button>
         </form>
         <div id="singer-list">
-
             <?php require_once "../data/mysql-connection.php";
             $query = "SELECT sg.singer_id, sg.singer_name, COUNT(s.s_id) as song_cnt
             FROM singers sg LEFT OUTER JOIN songs s ON s.singer_id = sg.singer_id 
@@ -98,19 +103,62 @@ require_once "../templates/header.php";
 </main>
 <script src="../js/ajax.js?v=<?php echo time() ?>"> </script>
 <script>
+    // edit song
+    document.querySelectorAll("#edit-song-list .edit").forEach(function(edit) {
+        edit.addEventListener("click", function(e) {
+            let songId = e.target.parentNode.attributes['editFor'].value;
+            let songNameElement = e.target.parentNode.previousElementSibling;
+            let songName = songNameElement.innerText;
+            if (songNameElement.outerText !== ' Update') {
+                songNameElement.parentElement.classList.add("flex-col");
+                songNameElement.innerHTML =
+                    `<input value='${songName}' oldValue='${songName}' />
+                     <button class='update-btn' onclick='updateSongName(this.parentNode,${songId});'>Update</button>
+                `;
+                songNameElement.firstChild.focus();
+            }
+        })
+    })
+
+    function updateSongName(e, songId) {
+        let newSongName = e.firstChild.value.trim();
+        let oldName = e.firstChild.attributes['oldValue'].value;
+        if (newSongName !== oldName && newSongName) {
+            updateSong(songId, newSongName, e);
+        } else {
+            e.innerHTML = `${oldName}`;
+        }
+        e.parentElement.classList.remove("flex-col");
+    }
+
+    // delete song
+    document.querySelectorAll("#edit-song-list .del").forEach(function(del) {
+        del.addEventListener("click", function(e) {
+            let parentNode = e.target.parentNode;
+            let songId = parentNode.attributes['editFor'].value;
+            if (parentNode.previousElementSibling.outerText !== ' Update') {
+                if (confirm("Do you really want to delete this song ?")) {
+                    deleteSong(songId, parentNode);
+                }
+            }
+        })
+    })
+
     // add new singer
     document.getElementById("singer-submit-btn").addEventListener("click", function(e) {
         e.preventDefault();
         document.getElementById("singer-warning").textContent = ""
-        let singer_name = document.getElementById("singer-name").value;
+        let singerName = document.getElementById("singer-name").value.trim();
         let selectSingers = document.getElementById("song-singer");
         let singerWarning = document.getElementById("singer-warning");
         let singerList = document.getElementById("singer-list");
         let lastIndex = singerList.lastElementChild.lastElementChild.attributes['editFor'].value;
-        if (!singer_name) {
-            textContent = "Please enter singer name"
+        if (!singerName) {
+            singerWarning.textContent = "Please enter singer name"
+            document.getElementById("singer-name").focus();
         } else {
-            addSinger(singer_name, selectSingers, singerList, singerWarning, lastIndex);
+            addSinger(singerName, selectSingers, singerList, singerWarning, lastIndex);
+            document.getElementById("singer-name").value = "";
         }
     });
 
@@ -125,15 +173,16 @@ require_once "../templates/header.php";
                     `<input value='${singerName}' oldValue='${singerName}' />
                     <button class='update-btn' onclick='updateSingerName(this.parentNode,${singerId});'>Update</button>
                 `;
+                singerNameElement.firstChild.focus();
             }
         })
     })
 
     function updateSingerName(e, singerId) {
-        let newSingerName = e.firstChild.value;
+        let newSingerName = e.firstChild.value.trim();
         let singerName = e.firstChild.attributes['oldValue'].value;
 
-        if (newSingerName !== singerName) {
+        if (newSingerName !== singerName && newSingerName) {
             updateSinger(singerId, newSingerName, e);
         } else {
             e.innerHTML = `${singerName}`;
@@ -148,15 +197,13 @@ require_once "../templates/header.php";
             let singerId = parentNode.attributes['editFor'].value;
             let numberSong = parentNode.attributes['numberSong'].value;
             if (parentNode.previousElementSibling.outerText !== ' Update' && numberSong === '0') {
-                deleteSinger(singerId, parentNode);
+
+                if (confirm("Do you really want to delete this singer ?")) {
+                    deleteSinger(singerId, parentNode);
+                }
             }
         })
     })
-
-    // function updateSingerName(e, singerId) {
-    //     let singerName = e.firstChild.value;
-    //     updateSinger(singerId, singerName, e);
-    // }
 </script>
 <?php
 require_once '../templates/footer.php'
